avm1: Move joint AVM1 AMF code to its own file
