+++ core/src/avm1/globals/local_connection.rs
@@ -1,8 +1,8 @@
 //! LocalConnection class
 
 use crate::avm1::activation::Activation;
+use crate::avm1::amf::{deserialize_value, serialize};
 use crate::avm1::error::Error;
-use crate::avm1::globals::shared_object::{deserialize_value, serialize};
 use crate::avm1::property_decl::{DeclContext, PropertyOrder, StaticDeclarations, SystemClass};
 use crate::avm1::{ActivationIdentifier, ExecutionReason, NativeObject, Object, Value};
 use crate::avm1_stub;
